Added Zod validation on server and client side

+++ src/actions/actions.ts
@@ -1,15 +1,25 @@
 "use server";
 
 import prisma from "@/lib/db";
+import { petFormSchema, petIdSchema } from "@/lib/schemas";
 import { PetEssentials } from "@/lib/types";
+import { sleep } from "@/lib/utils";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(pet: PetEssentials) {
-  await new Promise((resolve) => setTimeout(resolve, 2000));
+export async function addPet(pet: unknown) {
+  sleep(2000);
+
+  const validationResult = petFormSchema.safeParse(pet);
+  if (!validationResult.success) {
+    return {
+      success: false,
+      message: "Invalid form data",
+    };
+  }
 
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validationResult.data,
     });
   } catch (err) {
     return {
@@ -25,13 +35,31 @@ export async function addPet(pet: PetEssentials) {
   };
 }
 
-export async function editPet(petId: string, pet: PetEssentials) {
+export async function editPet(petId: unknown, pet: unknown) {
+  sleep(2000);
+
+  const validationResultId = petIdSchema.safeParse(petId);
+  if (!validationResultId.success) {
+    return {
+      success: false,
+      message: "Invalid pet ID",
+    };
+  }
+
+  const validationResultPet = petFormSchema.safeParse(pet);
+  if (!validationResultPet.success) {
+    return {
+      success: false,
+      message: "Invalid form data",
+    };
+  }
+
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validationResultId.data,
       },
-      data: pet,
+      data: validationResultPet.data,
     });
   } catch (err) {
     return {
@@ -47,11 +75,21 @@ export async function editPet(petId: string, pet: PetEssentials) {
   };
 }
 
-export async function deletePet(petId: string) {
+export async function deletePet(petId: unknown) {
+  sleep(2000);
+
+  const validationResult = petIdSchema.safeParse(petId);
+  if (!validationResult.success) {
+    return {
+      success: false,
+      message: "Invalid pet ID",
+    };
+  }
+
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validationResult.data,
       },
     });
   } catch (err) {
